+++ components/input/index.js
@@ -1,1 +1,2 @@
-export { Input } from './component';
+export { Input } from './Input';
+export { FilterInput } from './FilterInput';